add stricts to api

+++ internal/app/aim/usecases/interactor/deleteButton.go
@@ -2,6 +2,7 @@ package interactor
 
 import (
 	"context"
+	"fmt"
 
 	"github.com/NaKa2355/aim/internal/app/aim/entities/button"
 	remote "github.com/NaKa2355/aim/internal/app/aim/entities/remote"
@@ -9,9 +10,13 @@ import (
 )
 
 func (i *Interactor) deleteButton(ctx context.Context, in bdy.DeleteButtonInput) (err error) {
-	_, err = i.repo.ReadRemote(ctx, remote.ID(in.RemoteID))
+	b, err := i.repo.ReadButton(ctx, remote.ID(in.RemoteID), button.ID(in.ButtonID))
 	if err != nil {
 		return
+	}
+
+	if b.Tag != "" {
+		return bdy.NewError(bdy.CodeInvaildOperation, fmt.Errorf("cannot delete a button which has tag"))
 	}
 
 	err = i.repo.DeleteButton(ctx, remote.ID(in.RemoteID), button.ID(in.ButtonID))
